MVC Fix, BOARDSIZE Variable and Observer Pattern
Fixed MVC so that board has a controller referrence that it can call to notify view that it needs to update. Added the observer pattern with the board being observable and the view having the draw() method that is updated when handtiles or boardtiles changes. Added a final boardsize variable and removed hard coded instances of board size so that it can now be set by changing that variable.